# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: golang-clean API
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      summary: Liveness probe
      operationId: healthz
      responses:
        '200':
          description: Service is alive
  /readyz:
    get:
      summary: Readiness probe
      operationId: readyz
      responses:
        '200':
          description: Service is ready
        '503':
          description: Service is not ready
  /metrics:
    get:
      summary: Prometheus metrics endpoint
      operationId: metrics
      responses:
        '200':
          description: Prometheus metrics
  /api/v1/orders:
    post:
      summary: Create order
      operationId: createOrder
      security:
        - ApiKeyAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer_id, currency, amount]
              properties:
                customer_id:
                  type: string
                currency:
                  type: string
                amount:
                  type: integer
                  format: int64
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
    get:
      summary: List orders
      operationId: listOrders
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /api/v1/orders/{id}:
    get:
      summary: Get order by id
      operationId: getOrder
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
    put:
      summary: Update order
      operationId: updateOrder
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [customer_id, currency, amount]
              properties:
                customer_id:
                  type: string
                currency:
                  type: string
                amount:
                  type: integer
                  format: int64
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
    delete:
      summary: Delete order
      operationId: deleteOrder
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Order deleted
        '404':
          description: Order not found
  /api/v1/users:
    post:
      summary: Create user
      operationId: createUser
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          description: Email already exists
    get:
      summary: List users
      operationId: listUsers
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /api/v1/users/{id}:
    get:
      summary: Get user by id
      operationId: getUser
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      summary: Update user
      operationId: updateUser
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
        '409':
          description: Email already exists
    delete:
      summary: Delete user
      operationId: deleteUser
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted
        '404':
          description: User not found
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        type:
          type: string
        message:
          type: string
        request_id:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
        customer_id:
          type: string
        currency:
          type: string
        amount:
          type: integer
          format: int64
        status:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email